pyproject.toml
[project]
name = "searchless-chess"
version = "1.0.0"
description = "Searchless Chess — Neural position evaluation without tree search"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Antoni Czapski" }]

dependencies = [
    "torch>=2.1",
    "numpy>=1.24",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
    "loguru>=0.7",
    "python-chess>=1.999",
    "datasets>=2.14",
    "scikit-learn>=1.3",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
train = [
    "pytorch-lightning>=2.1",
    "wandb>=0.16",
    "torchmetrics>=1.2",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.4",
]
analysis = [
    "jupyter>=1.0",
    "matplotlib>=3.7",
    "seaborn>=0.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
